ajout du blockly dans lanalyse
git commit -am a

+++ LaRuche_2017/laruche/js_tools/blockly/core/workspace.js
@@ -271,9 +271,11 @@ Blockly.Workspace.prototype.createVariable = function(name) {
     this.variableList.push(name);
   }
   //AJOUT POUR LE PROJET LA RUCHE
-  if(!variable_List[name]){
-    variable_List[name] = new Variable(name,'Real');
-    update_all_view();
+  if(this.id == prepEditor.mBlocklyWorkspace.id){
+    if(!variable_List[name]){
+      variable_List[name] = new Variable(name,'Real');
+      update_all_view();
+    }
   }
 };
 
@@ -308,12 +310,14 @@ Blockly.Workspace.prototype.getVariableUses = function(name) {
  */
 Blockly.Workspace.prototype.deleteVariable = function(name) {
   //AJOUT
-  delete variable_List[name];
-  editor_Enonce.destroy_var(name);
-  for(var key in answer_List){
-    answer_List[key].get_block_html().get_editor().destroy_var(name);
+  if(this.id == prepEditor.mBlocklyWorkspace.id){
+    delete variable_List[name];
+    editor_Enonce.destroy_var(name);
+    for(var key in answer_List){
+      answer_List[key].get_block_html().get_editor().destroy_var(name);
+    }
+    update_all_view();
   }
-  update_all_view();
   //FIN DE L'AJOUT
   var variableIndex = this.variableIndexOf(name);
   if (variableIndex == -1) {
